Fix Next.js links not working
