+++ theme/src/utils/ssg.ts
@@ -4,10 +4,112 @@ import cheerio, { Cheerio, Element } from 'cheerio';
 import fs from 'fs-extra';
 
 import {
-  getGlobalHeaders,
+  GlobalHeader,
   JupyterBookState,
   TOCHeader,
 } from '@/context/jupyterBook';
+
+interface StackItem<T> {
+  level: number;
+  node: T;
+  parentId?: string;
+}
+
+interface LinkData {
+  href: string;
+  text: string;
+}
+
+interface GetGlobalHeadersOptions<T> {
+  /**
+   * Gets the initial list of TOC tree nodes to parse.
+   */
+  getRootNodes(): T[];
+
+  /**
+   * Gets link data for some link under the given parent node.
+   *
+   * @param parentNode The parent node.
+   */
+  getLinkData(parentNode: T): LinkData;
+
+  /**
+   * Gets the next level of TOC tree nodes given some parent node and next level.
+   *
+   * @param parentNode The parent node.
+   */
+  getNextNodes(parentNode: T, level: number): T[];
+}
+
+/**
+ * Utility function for getting global header data from a DOM tree. The function
+ * is implemented generically enough so that it can be used on the browser and
+ * by the pre-renderer. Since the DOM is a tree, we can use Depth First Search
+ * to traverse the tree and deserialize it into the global headers data
+ * structures.
+ */
+function getGlobalHeaders<T>({
+  getRootNodes,
+  getLinkData: getLinkNode,
+  getNextNodes,
+}: GetGlobalHeadersOptions<T>) {
+  const globalHeaders: Record<string, GlobalHeader> = {};
+  const rootGlobalHeaders: string[] = [];
+
+  const stack: StackItem<T>[] = getRootNodes().map((node) => ({
+    node,
+    level: 1,
+  }));
+
+  while (stack.length > 0) {
+    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+    const { level, node, parentId } = stack.pop()!;
+
+    const link = getLinkNode(node);
+
+    if (link.href && link.text) {
+      // Add new header mapped to the link URL.
+      globalHeaders[link.href] = {
+        level,
+        children: [],
+        href: link.href,
+        text: link.text,
+      };
+
+      if (level === 1) {
+        rootGlobalHeaders.push(link.href);
+      }
+
+      // Add current header as child of parent header.
+      if (parentId) {
+        globalHeaders[parentId].children.push(link.href);
+      }
+
+      const nextLevel = level + 1;
+      const nextNodes = getNextNodes(node, nextLevel);
+
+      // Add next headers to the stack for processing.
+      for (const nextNode of nextNodes) {
+        stack.push({
+          parentId: link.href,
+          level: level + 1,
+          node: nextNode,
+        });
+      }
+    }
+  }
+
+  // Reverse header lists because stacks pop items in reverse.
+  rootGlobalHeaders.reverse();
+  for (const header of Object.values(globalHeaders)) {
+    header.children.reverse();
+  }
+
+  return {
+    globalHeaders,
+    rootGlobalHeaders,
+  };
+}
 
 function getPageHeaders(contentToc: Cheerio<Element>) {
   return contentToc
